refactor: mudando form avaliação  de modal para página
