+++ src/templates/areaDeTrabalho/components/StudentInfoInput.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import styles from "../styles/FormularioAvaliacao.module.css";
+import styles from "../../../styles/formularioAvaliação/FormularioAvaliacao.module.css";
 
 const noop = () => null;
